Implemented GNU plots and spectrograms for S/X analysis
